+++ sdk/cosmos/azure-cosmos-spark_3-1_2-12/src/main/scala/com/azure/cosmos/spark/CosmosConstants.scala
@@ -14,7 +14,7 @@ private object CosmosConstants {
 
   object Names {
     val ItemsDataSourceShortName = "cosmos.oltp"
-    val ChangeFeedDataSourceShortName = "cosmos.changeFeed"
+    val ChangeFeedDataSourceShortName = "cosmos.oltp.changeFeed"
   }
 
   object Properties {